+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [3.6.0] - 2022-01-14
 ### Added
 - Budget Type: Static or Dynamic 
 - Static Budget Type implementation
@@ -85,7 +87,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [2.42.1] - 2021-12-22
 
-[Unreleased]: https://github.com/ArtemNikolaev/toLiveToPay/compare/v3.5.0...HEAD
+[Unreleased]: https://github.com/ArtemNikolaev/toLiveToPay/compare/v3.6.0...HEAD
+[3.6.0]: https://github.com/ArtemNikolaev/toLiveToPay/compare/v3.5.0...v3.6.0
 [3.5.0]: https://github.com/ArtemNikolaev/toLiveToPay/compare/v3.4.1...v3.5.0
 [3.4.1]: https://github.com/ArtemNikolaev/toLiveToPay/compare/v3.4.0...v3.4.1
 [3.4.0]: https://github.com/ArtemNikolaev/toLiveToPay/compare/v3.3.0...v3.4.0